% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-timer.r
\name{add_timer}
\alias{add_timer}
\title{Wrap a function to add a timer}
\usage{
add_timer(.f)
}
\arguments{
\item{.f}{A function to modify to add a timer.}
}
\value{
Wrapped function that returns a list with components `result` and
  `time`.
}
\description{
Modify a function so its execution is timed with `system.time()` and both
the results of the funciton call and the execution time are returned as a
list.
}
\examples{
mean_timed <- add_timer(mean)
mean_timed(rnorm(1e6))

lm_timed <- add_timer(lm)
x <- (1:1e6) / 1e6
y = 2 * x + 1 + rnorm(x)
lm_timed(y ~ x)
}
