% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-this.r
\name{time_this}
\alias{time_this}
\title{Time an expression or code chunk}
\usage{
time_this(code, time_as = c("list", "attr"))
}
\arguments{
\item{code}{An expression or code chunk (wrapped in curly brackets) to time.}

\item{time_as}{character; return the execution time and results as list
elements ("list") or return results with time as an attribute ("attr").}
}
\value{
A list with elements `result` and `time`, containing the results of
  evaluating `code` and the execution time, respectively. If `time_as =
  "attr"`, the result will be returned directly with the time stored as an
  attribute.
}
\description{
Time an expression or chunk of R code with `system.time()` and return both
the results of evaluating the expression and the execution time in a list.
}
\examples{
# time a simple expression
time_this(mean(rnorm(1e6)))

# time a code chunk
time_this({
  x <- (1:1e6) / 1e6
  y = 2 * x + 1 + rnorm(x)
  lm(y ~ x)
})

# time can also be stored as an attribute
time_this(mean(rnorm(1e6)), time_as = "attr")
}
